Split helpers and utils from actually.go to keep it tiny
