+++ actually.go
@@ -2,14 +2,9 @@
 package actually
 
 import (
-	"os"
-	"strings"
 	"testing"
 
-	"github.com/bayashi/actually/diff"
-	"github.com/bayashi/actually/report"
 	"github.com/bayashi/actually/testobject"
-	"github.com/bayashi/actually/trace"
 )
 
 // TestingA is a context of the test
@@ -63,143 +58,4 @@ func (a *TestingA) Expect(e any) *TestingA {
 	a.setExpect = true
 
 	return a
-}
-
-func failNowPtr(v bool) *bool {
-	return &v
-}
-
-// FailNotNow turns a flag so that even if the test fails, execution does not stop immediately.
-/*
-   It behaves this way by default. If you want the opposite behavior, call `FailNow` method.
-   NOTE that FailNotNow method should be called after `Got` or `Expect`.
-*/
-func (a *TestingA) FailNotNow() *TestingA {
-	a.failNow = failNowPtr(false)
-
-	return a
-}
-
-// FailNotNowOn function turns off an ENV flag to stop further test execution immediately if one test fails.
-/*
-	func Test(t *testing.T) {
-		actually.FailNowOn(t)
-		actually.Got(something).Nil(t)                    // Fail Now
-		actually.Got(something).Expect(something).Same(t) // Fail Now
-
-		actually.FailNotNowOn(t)
-		actually.Got(something).Nil(t)                    // NOT Fail Now
-		actually.Got(something).Expect(something).Same(t) // NOT Fail Now
-
-		actually.Got(something).FailNow().Nil(t)          // Fail Now
-	}
-*/
-func FailNotNowOn(t *testing.T) {
-	t.Setenv(envKey_FailNow, "")
-}
-
-// `FailNow` turns on a flag to stop further test execution immediately if one test fails
-// NOTE that FailNow method should be called after `Got` or `Expect`.
-/*
-	actually.Got(something).FailNow().Nil(t) // Fail now for only this test
-*/
-func (a *TestingA) FailNow() *TestingA {
-	a.failNow = failNowPtr(true)
-
-	return a
-}
-
-// FailNowOn function turns on an ENV flag to stop further test execution immediately if one test fails.
-/*
-	func Test(t *testing.T) {
-		actually.FailNowOn(t)
-		actually.Got(something).Nil(t)                    // Fail Now
-		actually.Got(something).Expect(something).Same(t) // Fail Now
-	}
-*/
-func FailNowOn(t *testing.T) {
-	t.Setenv(envKey_FailNow, envKey_FailNow)
-}
-
-func (a *TestingA) fail(r *report.Report) *TestingA {
-	a.t.Helper()
-	r.Trace(traceinfo()).Function(a.t.Name() + "()").Name(a.name)
-	a.t.Errorf("\n%s", r.Put())
-	if a.failNow != nil && !*a.failNow {
-		a.t.Fail()
-	} else if (a.failNow != nil && *a.failNow) || len(os.Getenv(envKey_FailNow)) > 0 {
-		a.t.FailNow()
-	} else {
-		a.t.Fail()
-	}
-
-	return a
-}
-
-// X turns on a flag to show test values as raw in a fail report.
-func (a *TestingA) X() *TestingA {
-	a.showRawData = true
-
-	return a
-}
-
-var skipTraceRule = func(filepath string) bool {
-	// Skip myself
-	return strings.Contains(filepath, "actually.go")
-}
-
-func traceinfo() string {
-	return strings.Join(trace.Info(skipTraceRule), traceSeparator)
-}
-
-// Skip provides shorthand to skip further tests within the same function for `-short` option.
-/*
-	func Test(t *testing.T) {
-		Got(1).NotNil(t) // Run
-		Skip(t)
-		Got(2).NotNil(t) // Skip
-		Got(3).NotNil(t) // Skip Also
-	}
-*/
-func Skip(t *testing.T, skipReasons ...any) {
-	if testing.Short() {
-		t.Skip(skipReasons...)
-	}
-}
-
-// Set test name spefically.
-/*
-	You can also set a test name on assertion methods.
-
-	actually.Got(a).Expect(b).Same(t, "Test Name")
-*/
-func (a *TestingA) Name(n string) *TestingA {
-	a.name = n
-
-	return a
-}
-
-func (a *TestingA) naming(testNames ...string) string {
-	if a.name != "" {
-		if len(testNames) > 0 {
-			n := []string{a.name}
-			n = append(n, testNames...)
-			return strings.Join(n, ", ")
-		} else {
-			return a.name
-		}
-	} else {
-		return strings.Join(testNames, ", ")
-	}
-}
-
-// Diff is a helper function to get a diff string of 2 objects for debugging
-func Diff(a any, b any) string {
-	return diff.DiffSimple(a, b)
-}
-
-func invalidCall(a *TestingA) {
-	if !a.setGot {
-		panic("You called assertion method, but you forgot to call Got().")
-	}
 }
